Moving Redux store creation to index.js within /Reducers
