+++ client/src/Reducers/index.js
@@ -1,3 +1,8 @@
+import { createStore, applyMiddleware, combineReducers, compose } from 'redux';
+
+// Middlewares
+import thunk from 'redux-thunk';
+
 import { BalanceReducer } from './BalanceReducer';
 import { DepositReducer } from './DepositReducer';
 import { LoginReducer } from './LoginReducer';
@@ -9,8 +14,7 @@ import { TransactionReducer } from './TransactionReducer';
 import { UserReducer } from './UserReducer';
 import { WithdrawalReducer } from './WithdrawalReducer';
 
-
-export {
+const rootReducer = combineReducers({
   BalanceReducer,
   DepositReducer,
   LoginReducer,
@@ -21,4 +25,12 @@ export {
   TransactionReducer,
   UserReducer,
   WithdrawalReducer,
-};
+});
+
+export const store = createStore(
+  rootReducer, {},
+  compose(
+    applyMiddleware(thunk),
+    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+  )
+);
